Implement Checking Running Process Service 현재 실행 중인 프로세스를 검사하는 서비스 구현
