+++ app/src/main/AndroidManifest.xml
@@ -13,13 +13,13 @@
         android:theme="@style/AppTheme"
         tools:replace="android:icon">
         <activity android:name=".presenter.Activity.MainActivity">
-
-        </activity>
-        <activity android:name=".presenter.Activity.LoginActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity android:name=".presenter.Activity.LoginActivity">
+
         </activity>
         <activity android:name=".presenter.LockActivity">
         </activity>
