+++ app/src/main/java/com/example/ready/studytimemanagement/presenter/AppLockService.java
@@ -1,12 +1,63 @@
 package com.example.ready.studytimemanagement.presenter;
 
+import android.app.AppOpsManager;
 import android.app.Service;
+import android.content.Context;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.os.IBinder;
 import android.support.annotation.Nullable;
 import android.util.Log;
 
 public class AppLockService extends Service {
+
+    AppLockController alc;
+    checkThread th;
+    private Context context = null;
+
+    boolean checkFlag;
+
+    private class checkThread extends Thread{
+        public void run() {
+            int num = 1;
+
+            // GET_USAGE_STATS 권한 확인
+            boolean granted = false;
+            AppOpsManager appOps = (AppOpsManager) context.getSystemService(Context.APP_OPS_SERVICE);
+            int mode = appOps.checkOpNoThrow(AppOpsManager.OPSTR_GET_USAGE_STATS,android.os.Process.myUid(), context.getPackageName());
+
+            if (mode == AppOpsManager.MODE_DEFAULT) {
+                granted = (context.checkCallingOrSelfPermission(android.Manifest.permission.PACKAGE_USAGE_STATS) == PackageManager.PERMISSION_GRANTED);
+            } else {
+                granted = (mode == AppOpsManager.MODE_ALLOWED);
+            }
+
+            Log.d("isRooting granted = " , String.valueOf(granted));
+
+            if (granted == false)
+            {
+                // 권한이 없을 경우 권한 요구 페이지 이동
+                Intent intent = new Intent(android.provider.Settings.ACTION_USAGE_ACCESS_SETTINGS);
+                context.startActivity(intent);
+            }
+
+            while(checkFlag) {
+                if(alc.CheckRunningApp(context)) {
+                    Intent intent = new Intent(context,LockActivity.class);
+                    intent.setFlags(Intent.FLAG_ACTIVITY_NEW_TASK);
+                    startActivity(intent);
+                    //finish();
+                }
+                Log.d("Thread", "" + num++);
+                try {
+                    Thread.sleep(1000 );
+                } catch (InterruptedException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+    }
+
     @Nullable
     @Override
     public IBinder onBind(Intent intent) {
@@ -19,11 +70,24 @@ public class AppLockService extends Service {
         // 서비스에서 가장 먼저 호출됨(최초에 한번만)
         Log.d("Service : ", "서비스의 onCreate");
 
+        alc = new AppLockController();
+        checkFlag = false;
+        context = getApplicationContext();
     }
     @Override
     public int onStartCommand(Intent intent, int flags, int startId) {
         // 서비스가 호출될 때마다 실행
-        Log.d("Service : ", "서비스의 onStartCommand");
+        Log.d("Service : ", "서비스의 onStartCommand - "+flags+"번 서비스");
+
+
+
+        if(checkFlag==false) {
+            th = new checkThread();
+            th.start();
+        }else {
+            stopSelf();
+        }
+        checkFlag = !checkFlag;
 
         return super.onStartCommand(intent, flags, startId);
     }
@@ -31,6 +95,8 @@ public class AppLockService extends Service {
     @Override
     public void onDestroy() {
         super.onDestroy();
+        th = null;
+        Log.d("Thread", "쓰레드 뿌셔");
         // 서비스가 종료될 때 실행
 
         Log.d("Service : ", "서비스의 onDestroy");
